% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{as.mesh3d.fracture_geom}
\alias{as.mesh3d.fracture_geom}
\title{Makes a mesh3d object from fracture}
\usage{
as.mesh3d.fracture_geom(
  obj,
  type = c("top", "bottom"),
  transform = function(points) points
)
}
\arguments{
\item{obj}{the fracture_geom object to plot in 3d}

\item{type}{the surface to plot: top, bottom, middle}

\item{...}{other arguments passed to tmesh3d}
}
\description{
Makes a mesh3d object from fracture
}
