% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_matrix.R
\name{expand_seq}
\alias{expand_seq}
\title{Expands a grid of sequences (outputs a matrix)}
\usage{
expand_seq(x, fun = seq_len, matrix = TRUE)
}
\description{
Expands a grid of sequences (outputs a matrix)
}
