% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balls.R
\name{fracture.balls}
\alias{fracture.balls}
\title{Generate balls inside of a 3d fracture}
\usage{
fracture.balls(
  obj,
  K = 4000,
  Rmax = 0.01,
  Rmin = 0.01,
  B = data.frame(),
  margin = 0.003,
  margin_opt = margin/3,
  relax_iterations = 10,
  max_add = 200,
  overshoot = 100,
  seed,
  mean.neighbor = 5,
  iterations = ceiling(1.5 * K/max_add),
  dist = function(k, Rmin, Rmax) runif(k, Rmin, Rmax),
  delete = TRUE,
  period = 1
)
}
\arguments{
\item{obj}{fracture_geom object}

\item{K}{target number of balls}

\item{Rmax, Rmin}{the limits on radiuses of sizes}

\item{B}{set of balls to start from}

\item{margin}{target margin distance between balls}

\item{margin_opt}{lower limit on margin for optimization}

\item{relax_iterations}{number of subiterations for relaxation}

\item{max_add}{number of balls added in each iteration}

\item{overshoot}{number of balls by which to overshoot target}

\item{seed}{random seed}

\item{mean.neighbor}{the mean.neighbor buffor size}

\item{iterations}{number of iterations}

\item{dist}{distribution of radii}

\item{delete}{if TRUE, allow the algorithm to delete balls during optimization}

\item{period}{the periodicity of the domain}
}
\description{
Generate balls inside of a 3d fracture
}
