% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam_geom.R
\name{fracture.geom.brown}
\alias{fracture.geom.brown}
\title{Generate a fracture generating function}
\usage{
fracture.geom.brown(
  scale = 1,
  alpha = 2,
  corr.profile,
  closed = 0.1,
  gap = NA,
  beta.make = FALSE,
  beta.shape = 4
)
}
\arguments{
\item{scale}{linear scale of the fracture}

\item{alpha}{exponent in the power spectrum}

\item{corr.profile}{correlation profile}

\item{closed}{the probability of touching walls}

\item{gap}{the mean distance between walls}

\item{beta.make}{if TRUE, transform into beta distribution}

\item{beta.shape}{beta distribution shape parameter (for beta.make=TRUE)}
}
\description{
Generate a fracture generating function
}
\references{
Brown, S. R. (1995). Simple mathematical model of a rough fracture. Journal of Geophysical Research: Solid Earth, 100(B4), 5941-5952.
}
