% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{fracture3d}
\alias{fracture3d}
\title{Makes a 3d plot of the fracture}
\usage{
fracture3d(
  obj,
  type = c("top", "bottom"),
  top = "top" \%in\% type,
  bottom = "bottom" \%in\% type,
  middle = "middle" \%in\% type,
  col = c(2, 3, 4),
  add = FALSE
)
}
\description{
Makes a 3d plot of the fracture
}
