% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_matrix.R
\name{matern}
\alias{matern}
\alias{matern.f}
\title{Calculate Matern covariance function and its power spectrum}
\usage{
matern(r, nu, l = 1)

matern.f(s, nu, l = 1)
}
\arguments{
\item{r}{distance}

\item{nu}{smoothness parameter}

\item{l}{scale}
}
\description{
Calculate Matern covariance function and its power spectrum
}
