% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.profiles.R
\name{ogilvie.corr.profile}
\alias{ogilvie.corr.profile}
\title{Correlation profile with linear cut-off}
\usage{
ogilvie.corr.profile(ML = 0.5, TL = 0, MinMF = 0, MaxMF = 1)
}
\arguments{
\item{ML}{Length of correlation cut-off}

\item{TL}{Width of linear cut-off fade}

\item{MinMF}{Minimal correlation}

\item{MaxMF}{Maximal correlation}
}
\description{
Correlation profile with linear cut-off
}
\references{
Steven R. Ogilvie, Evgeny Isakov, Paul W.J. Glover (2006), Fluid flow through rough fractures in rocks. II: A new matching model for rough rock fractures,
Earth and Planetary Science Letters, Volume 241, Issues 3–4, https://doi.org/10.1016/j.epsl.2005.11.041.
}
