% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam_geom.R
\name{ordered_rnorm_mat}
\alias{ordered_rnorm_mat}
\title{Generate a matrix of random complex numbers in a consistent order}
\usage{
ordered_rnorm_mat(N, M, k = 1, seed, length_one = FALSE)
}
\arguments{
\item{N}{number of rows}

\item{M}{number of columns}

\item{k}{number of independent matrices}

\item{seed}{random seed}

\item{length_one}{if TRUE, return complex numbers of length 1}
}
\description{
Generate a matrix of random complex numbers in a consistent order
}
