% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lammps.R
\name{read.lammps.data}
\alias{read.lammps.data}
\alias{write.lammps.data}
\title{Read and write balls from and to lammps data files}
\usage{
read.lammps.data(filename)

write.lammps.data(
  B,
  filename,
  density = 2,
  comment = "R generated balls",
  xlim = range(B$x),
  ylim = range(B$y),
  zlim = range(B$z)
)
}
\arguments{
\item{filename}{file path}

\item{B}{balls (data.frame)}

\item{density}{density of balls exported}

\item{comment}{comment placed in the header}

\item{xlim, ylim, zlim}{extent of the domain exported in the lammps file}
}
\description{
Read and write balls from and to lammps data files
}
