% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtri.R
\name{rtri}
\alias{rtri}
\title{Linear distribution over triangle}
\usage{
rtri(n, h1, h2, h3)
}
\arguments{
\item{n}{number of points to generate}

\item{h1, h2, h3}{weights in corners of triangle}
}
\description{
Takes a sample from a distribution which is a linear function over a triangle
}
\examples{
ret = rtri(1000, 1, 0, 0)
plot(ret$w1, ret$w2)

}
