% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_matrix.R
\name{seq_circ}
\alias{seq_circ}
\title{Generates a circular sequence of frequencies (0 to n/2 and -n/2 to 0)}
\usage{
seq_circ(n)
}
\arguments{
\item{n}{length of the sequence}
}
\description{
Generates a circular sequence of frequencies (0 to n/2 and -n/2 to 0)
}
