% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_geom.R
\name{set_gap}
\alias{set_gap}
\title{Set the gap and offset for already created fracture}
\usage{
set_gap(obj, gap, closed, offset)
}
\arguments{
\item{obj}{the fracture_geom object}

\item{gap}{the new gap}

\item{closed}{the probability of closed gap}

\item{offset}{the offset by which the fracture should be moved vertical}
}
\description{
Set the gap and offset for already created fracture
}
