% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Slices the fracture geometry}
\usage{
slice(
  obj,
  by = "h",
  flatten = "edge",
  value = "all",
  cut.triangles = TRUE,
  cut.edges = TRUE,
  cut.vertexes = TRUE,
  level = 0,
  eps = 1e-10,
  verbose = FALSE
)
}
\arguments{
\item{obj}{fracture_geom object}

\item{by}{name of the property/field by which to slice}

\item{flatten}{if to snap values to level}

\item{value}{type of the returned object (see return value)}

\item{level}{threshold level on the property}

\item{eps}{the numerical accuracy}

\item{verbose}{print information about cutting}
}
\value{
If value is "all", the function returns the same fracture_geom object, with some of the triangles sliced.
If value is "above", the function returns only the part of the geometry, for which property > level
"below" returns the opposite.
"equal" returns a list(points, edges), where points are points with property == level, and edges are the connecting edges (two columns)
}
\description{
Slices the fracture with respect to a specific property, and threshold level
}
\examples{
power.iso = function(f) 0.0001 * f^{-4.5}
ret = fracture_geom(refine=4, power.iso = power.iso)
ret2 = slice(ret, value="above")

}
