% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reynolds.R
\name{solve_cg}
\alias{solve_cg}
\title{Simple Conjugate Gradient implementation}
\usage{
solve_cg(
  A,
  b,
  itmax = 800,
  epsjump = 1e-06,
  eps,
  x0 = rep(0, ncol(A)),
  print.every = 500
)
}
\arguments{
\item{A}{Matrix}

\item{b}{Right-hand-side}

\item{itmax}{Maximum number of iterations}

\item{epsjump}{The level of residual drop needed}

\item{eps}{Absolute level of residual needed, by default equal to epsjump times initial residual}

\item{x0}{Initial guess on solution}
}
\description{
Simple Conjugate Gradient implementation
}
