% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reynolds.R
\name{solve_reynolds}
\alias{solve_reynolds}
\title{Calculated the premability tensor by solving a Reynolds lubrication equation in the fracture}
\usage{
solve_reynolds(obj, q.fun = function(h) h^3/12)
}
\arguments{
\item{obj}{the fracture_geom object}

\item{q.fun}{the dependence of the flux on the heights, by default h^3/12}
}
\description{
Calculated the premability tensor by solving a Reynolds lubrication equation in the fracture
}
