% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_geom.R
\name{surface_summary}
\alias{surface_summary}
\title{Summary of surfaces}
\usage{
surface_summary(obj, surface = c("f1", "f2", "h", "fm"))
}
\arguments{
\item{obj}{fracture_geom object}

\item{surface}{the variables over which to gather statistics}
}
\description{
Summary of surfaces
}
