% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_geom.R
\name{touching}
\alias{touching}
\title{Extract the overlapping part of two surfaces of a fracture}
\usage{
touching(obj, touch = "exclude")
}
\arguments{
\item{obj}{the fracture_geom object}

\item{touch}{what to do with the touching part (see description)}
}
\description{
The possible values of touch are: "exclude" - exclude touching part, "include" - leave untouched, "only" - select only the touching part.
}
