% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_geom.R
\name{volume.fracture_geom}
\alias{volume.fracture_geom}
\title{Calculate the volume of fracture geometry}
\usage{
\method{volume}{fracture_geom}(x, ...)
}
\arguments{
\item{x}{the fracture_geom object to calculate volume of}

\item{...}{other parameters (ignored)}
}
\description{
Calculate the volume of fracture geometry
}
