% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{volume}
\alias{volume}
\title{Generic function for volume calculation}
\usage{
volume(x, ...)
}
\arguments{
\item{x}{object of which to calculate volume}

\item{...}{other parameters}
}
\description{
Generic function for volume calculation
}
