% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{write.stl}
\alias{write.stl}
\alias{write.stl.default}
\title{Writes an stl file}
\usage{
write.stl(x, ...)

\method{write.stl}{default}(x, con, ascii = FALSE)
}
\arguments{
\item{x}{list of meshes to write}

\item{con}{file connection of filename to write to}

\item{ascii}{if TRUE, write in ASCII format (discouraged)}
}
\description{
Writes an stl file

Writes a mesh3d object (or a list) to an stl file
}
