% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vtk.R
\name{write.vtk.tet.mesh}
\alias{write.vtk.tet.mesh}
\title{Writes a tet-mesh of the fracture geometry in a legacy VTK format}
\usage{
write.vtk.tet.mesh(obj, filename, ...)
}
\arguments{
\item{obj}{the fracture_geom object}

\item{filename}{name of the VTK file to write}

\item{...}{arguments passed to extract.tet.mesh, such as transform, or genie.h}
}
\description{
Writes a tet-mesh of the fracture geometry in a legacy VTK format
}
