% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{cut.fracture_geom}
\alias{cut.fracture_geom}
\title{Cut the fracture geometry to a box}
\usage{
\method{cut}{fracture_geom}(x, ...)
}
\arguments{
\item{x}{fracture_geom object}

\item{...}{other arguments for slice}
}
\description{
Cut the fracture geometry to a box
}
