% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_geom.R
\name{cut.fracture_geom}
\alias{cut.fracture_geom}
\title{Cut the fracture geometry to a box}
\usage{
\method{cut}{fracture_geom}(x, eps = 1e-09, ...)
}
\arguments{
\item{x}{fracture_geom object}

\item{eps}{numerical margin}

\item{...}{other arguments}
}
\description{
Cut the fracture geometry to a box
}
