% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrums.R
\name{exp_spectrum}
\alias{exp_spectrum}
\title{Expotential spectrum}
\usage{
exp_spectrum(scale = 1, alpha = 2, length = 1)
}
\arguments{
\item{scale}{The scale factor}

\item{alpha}{The exponent}

\item{length}{The reference wavelength}
}
\description{
Expotential spectrum
}
