% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{fracture3d}
\alias{fracture3d}
\title{Makes a 3d plot of the fracture}
\usage{
fracture3d(
  obj,
  type = c("top", "bottom"),
  col,
  edge.col = NA,
  vertex.col = NA,
  asp = "iso",
  add = FALSE,
  transform = function(points) points,
  ...
)
}
\arguments{
\item{obj}{the fracture_geom object to plot in 3d}

\item{type}{the surface to plot: top, bottom, middle}

\item{col}{vector of 3 colors used for top, bottom and middle surface}

\item{add}{if TRUE, add plot to existing rgl window}
}
\description{
Makes a 3d plot of the fracture
}
