% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_matrix.R
\name{fracture_matrix}
\alias{fracture_matrix}
\title{Generate a fracture field matrix}
\usage{
fracture_matrix(
  dims = c(10, 10),
  span = diag(nrow = length(dims)),
  period = diag(nrow = length(dims)),
  power.spectrum = exp.spectrum(),
  corr.profile = function(k) 0,
  closed = 0.1,
  gap,
  seed,
  length_one = FALSE
)
}
\arguments{
\item{dims}{Dimensions of the fracture matrix (length N)}

\item{span}{base spanning the fracture (matrix NxN)}

\item{period}{base of periodicity parallelogram}

\item{corr.profile}{correlation profile between (function of wave length)}

\item{closed}{the probability of fields touching}

\item{gap}{mean gap between upper and lower fields (overwrites closed)}

\item{seed}{random seed}

\item{length_one}{use complex numbers of length one for field generation}

\item{spectrum}{power spectrum of the fields (function of frequency)}

\item{bonds}{upper and lower bonds for centering}

\item{cut}{if TRUE, the fields will be cut to bonds}

\item{widen}{the level of widen effect (see Description)}

\item{widen_grad}{the inclination of the widen effect}
}
\description{
Generate a fracture field matrix
}
