% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture_matrix.R
\name{fracture_matrix}
\alias{fracture_matrix}
\title{Generate a fracture field matrix}
\usage{
fracture_matrix(
  dims = c(10, 10),
  span = diag(nrow = length(dims)),
  period = diag(nrow = length(dims)),
  power.iso,
  power.spectrum = function(f) power.iso(sqrt(rowSums(f * f))),
  corr.profile = function(k) 0,
  closed = 0.1,
  gap,
  seed,
  corr.method = c("midline", "top", "mixed"),
  length_one = FALSE,
  gauss.order = 1
)
}
\arguments{
\item{dims}{Dimensions of the fracture matrix (length N)}

\item{span}{base spanning the fracture (matrix NxN)}

\item{period}{base of periodicity parallelogram}

\item{power.spectrum}{power spectrum of the fields (function of frequency)}

\item{corr.profile}{correlation profile between (function of wave length)}

\item{closed}{the probability of fields touching}

\item{gap}{mean gap between upper and lower fields (overwrites closed)}

\item{seed}{random seed}

\item{length_one}{use complex numbers of length one for field generation}
}
\description{
Generate a fracture field matrix
}
\examples{
seed = 123
power.spectrum = function(f) f^{-2.5}
par(mfrow=c(2,2))
ret = fracture_matrix(dims=c(50,50),span = diag(2), power.iso = power.spectrum, seed=seed)
plot(ret)
ret = fracture_matrix(dims=c(50,50),span = matrix(c(1,1,-1,1),2,2), power.iso = power.spectrum, seed=seed)
plot(ret)
ret = fracture_matrix(dims=c(50,50),span = diag(2)*2, power.iso = power.spectrum, seed=seed)
plot(ret)
ret = fracture_matrix(dims=c(50,50),span = diag(2)*2,period=diag(2)*2, power.iso = power.spectrum, seed=seed)
plot(ret)

}
\references{
Brown, S. R. (1995). Simple mathematical model of a rough fracture. Journal of Geophysical Research: Solid Earth, 100(B4), 5941-5952.
}
