% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.fracture_matrix}
\alias{plot.fracture_matrix}
\title{Plot fracture matrix}
\usage{
\method{plot}{fracture_matrix}(
  obj,
  field = "f1",
  col.palette = c("black", "red", "yellow"),
  pch = 16,
  cex = 1,
  asp = 1,
  ...
)
}
\arguments{
\item{obj}{fracture matrix to plot}

\item{field}{field to plot ("f1" - upper, "f2" - lower)}

\item{col.palette}{color palette to use for plotting}

\item{...}{other options for plot (scatter plot)}
}
\description{
Plot fracture matrix
}
