% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Slices the fracture geometry}
\usage{
slice(obj, by = "h", flatten = "edge", value = "all", level = 0, eps = 1e-10)
}
\arguments{
\item{obj}{fracture_geom object}

\item{by}{name of the property/field by which to slice}

\item{flatten}{if to snap values to level}

\item{value}{type of the returned object (see return value)}

\item{level}{threshold level on the property}

\item{eps}{the numerical accuracy}
}
\value{
If value is "all", the function returns the same fracture_geom object, with some of the triangles sliced.
If value is "above", the function returns only the part of the geometry, for which property > level
"below" returns the opposite.
"equal" returns a list(points, edges), where points are points with property == level, and edges are the connecting edges (two columns)
}
\description{
Slices the fracture with respect to a specific property, and threshold level
}
\examples{
library(rgl)
ret = fracture_geom(refine=4)
ret = slice(ret, eps=1e-2, flatten="above")
clear3d()
wire3d(as.mesh3d(ret))

}
